requirements.txt
neo4j-driver==1.1.0
neomodel==3.2.2
django_neomodel==0.0.3
django-cors-headers==2.0.2
django-extensions==1.7.8
django-fernet-fields==0.5
requests==2.13.0
numpy==1.12.1
sklearn==0.0
scikit-learn==0.18.1
scipy==0.19.0
iteration-utilities==0.6.1